refactor(hero): remove .module.css and excess JSDocs
