+++ src/app/page.tsx
@@ -4,7 +4,7 @@ import Footer from "@/components/layout/Footer";
 import NavBar from "@/components/layout/Navbar";
 import AnnouncementSection from "@/components/sections/announcement/AnnouncementSection";
 import EventsSection from "@/components/sections/events/EventsSection";
-import HeroSection from "@/components/sections/home/HomeSection";
+import HeroSection from "@/components/sections/hero/HeroSection";
 
 export default function App() {
   return (
