+++ src/components/ui/ScrollDownArrow.tsx
@@ -2,17 +2,12 @@ import React from "react";
 import Image from "next/image";
 
 /**
- * ScrollDownArrow
- * 
  * A centered, animated arrow-down icon for indicating scroll.
- * 
- * Usage:
- * <ScrollDownArrow />
  */
 const ScrollDownArrow: React.FC = () => (
     <div className="flex justify-center">
         <Image
-            src="/images/arrow-down-icon.svg"
+            src="/images/hero/arrow-down-icon.svg"
             alt="Scroll Down"
             width={50}
             height={50}
